Fået rabat til at virke

+++ Scooterland/Shared/Models/SalesLineItem.cs
@@ -14,7 +14,7 @@ namespace Scooterland.Shared.Models
 		public int ProductId { get; set; }
 		public Product? Product { get; set; }
 		public int Quantity { get; set; }
-		public decimal Discount { get; set; }
+		public int Discount { get; set; }
 
 		//Tom contructor til EF
 		public SalesLineItem()
